Use TypeScript and multi-package structure.
- CLI
- Lib
- VSCode Extension
