+++ .vscode/launch.json
@@ -1,0 +1,28 @@
+// A launch configuration that compiles the extension and then opens it inside a new window
+// Use IntelliSense to learn about possible attributes.
+// Hover to view descriptions of existing attributes.
+// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+{
+	"version": "0.2.0",
+	"configurations": [
+		{
+			"name": "Run VSCode Extension",
+			"type": "extensionHost",
+			"request": "launch",
+			"runtimeExecutable": "${execPath}",
+			"args": ["--extensionDevelopmentPath=${workspaceFolder}/plugins/vscode"],
+			"outFiles": ["${workspaceFolder}/plugins/vscode/out/*.js"],
+		},
+		{
+			"name": "VSCode Extension Tests",
+			"type": "extensionHost",
+			"request": "launch",
+			"runtimeExecutable": "${execPath}",
+			"args": [
+				"--extensionDevelopmentPath=${workspaceFolder}/plugins/vscode",
+				"--extensionTestsPath=${workspaceFolder}/plugins/vscode/out/test"
+			],
+			"outFiles": ["${workspaceFolder}/plugins/vscode/out/test/*.js"]
+		}
+	]
+}
