+++ cli/package.json
@@ -1,0 +1,38 @@
+{
+	"name": "nuget-extensions",
+	"version": "0.0.3-alpha",
+	"description": "Extensions for the NuGet CLI",
+	"main": "index.js",
+	"bin": {
+		"nuget-extensions": "./out/src/index.js"
+	},
+	"files": [
+		"out/**/*.js"
+	],
+	"scripts": {
+		"clean": "rimraf out",
+		"lint": "tslint -p ./",
+		"compile": "tsc -p ./",
+		"build": "yarn clean && yarn lint && yarn compile",
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"author": "Matt Preucil <mattpsbs@gmail.com>",
+	"license": "MIT",
+	"bugs": {
+		"url": "https://github.com/randymarsh77/nuget-extensions/issues"
+	},
+	"homepage": "https://github.com/randymarsh77/nuget-extensions#readme",
+	"devDependencies": {
+		"@types/node": "^11.10.4",
+		"@types/yargs": "^12.0.9",
+		"prettier": "^1.16.4",
+		"tslint": "^5.13.1",
+		"typescript": "^3.3.3333"
+	},
+	"dependencies": {
+		"lodash.debounce": "^4.0.8",
+		"rimraf": "^2.6.3",
+		"user-home": "^2.0.0",
+		"yargs": "^13.2.1"
+	}
+}
