+++ cli/src/index.ts
@@ -1,20 +1,19 @@
 #!/usr/bin/env node
 
-const { link } = require('./linker');
-const { registerPackages, readRegistry, writeRegistry } = require('./registry');
-const { watch } = require('./watcher');
+import * as yargs from 'yargs';
+import { link, registerPackages, readRegistry, writeRegistry, watch } from '../../lib/src';
 
-require('yargs')
+const _ = yargs
 	.usage('Usage: $0 <command> [options]')
 	.command(
 		'link [projects]',
 		'Link some packages. Passing no arguments will register packages in the current directory as link targets.',
-		yargs => {
-			yargs.positional('projects', {
+		command =>
+			command.positional('projects', {
+				type: 'string',
 				describe:
 					'If specified, links all registered packages to the specified project file(s). Specify multiple projects using a comma separated list.',
-			});
-		},
+			}),
 		argv => {
 			if (argv.projects) {
 				link(argv.projects.split(','));
@@ -23,22 +22,26 @@ require('yargs')
 			}
 		}
 	)
-	.command('reset', 'Reset the registry.', () => {
-		writeRegistry({});
+	.command('reset', 'Reset the registry.', {
+		handler: () => {
+			writeRegistry({});
+		},
 	})
-	.command('list', 'List linked packages.', () => {
-		console.log(JSON.stringify(readRegistry(), null, 2));
+	.command('list', 'List linked packages.', {
+		handler: () => console.log(JSON.stringify(readRegistry(), null, 2)),
 	})
 	.command(
 		'watch',
 		'Watch for file changes in linked package directories, automatically re-install the updated packages.',
-		yargs => {
-			yargs.option('short-circuit-build', {
+		command =>
+			command.option('short-circuit-build', {
+				type: 'string',
 				describe:
 					'If specified, the automatically re-installed package dlls will also be copied to the specified location.',
-			});
-		},
+			}),
 		argv => {
-			watch(argv);
+			watch({
+				shortCircuitBuild: argv['short-circuit-build'],
+			});
 		}
 	).argv;
