+++ lib/src/index.ts
@@ -1,0 +1,4 @@
+export * from './csproj';
+export * from './linker';
+export * from './registry';
+export * from './watcher';
