+++ lib/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../tsconfig.shared.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"outDir": "out"
+	}
+}
