+++ package.json
@@ -1,30 +1,7 @@
 {
-  "name": "nuget-extensions",
-  "version": "0.0.1-alpha",
-  "description": "Extensions for the NuGet CLI",
-  "main": "index.js",
-  "bin": {
-    "nuget-extensions": "./index.js"
-  },
-  "files": [
-    "*.js"
-  ],
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Matt Preucil <mattpsbs@gmail.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/randymarsh77/nuget-extensions/issues"
-  },
-  "homepage": "https://github.com/randymarsh77/nuget-extensions#readme",
   "devDependencies": {
-    "prettier": "^1.16.4"
-  },
-  "dependencies": {
-    "lodash.debounce": "^4.0.8",
-    "rimraf": "^2.6.3",
-    "user-home": "^2.0.0",
-    "yargs": "^13.2.1"
+    "@types/lodash.debounce": "^4.0.5",
+    "@types/rimraf": "^2.0.2",
+    "@types/user-home": "^2.0.0"
   }
 }
