+++ plugins/vscode/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../tsconfig.shared.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"outDir": "out"
+	},
+	"exclude": ["node_modules", ".vscode-test"]
+}
